Move GrafikElement form adapter to data layer
